+++ rodchenkov.simon/A1/circle.hpp
@@ -7,17 +7,17 @@ class Circle final : public Shape
 {
 public:
   Circle();
-  Circle(const point_t& pos, const float radius);
+  Circle(const point_t& pos, const double radius);
 
-  float       getArea() const noexcept override;
+  double       getArea() const noexcept override;
   rectangle_t getFrameRect() const noexcept override;
   void        printData(std::ostream& stream) const override;
 
   void move(const point_t& point) noexcept override;
-  void move(const float   dx, const float dy) noexcept override;
+  void move(const double   dx, const double dy) noexcept override;
 
 private:
-  float       radius_;
+  double       radius_;
   rectangle_t frame_rect_;
 };
