+++ rodchenkov.simon/A1/rectangle.hpp
@@ -7,18 +7,18 @@ class Rectangle final : public Shape
 {
 public:
   Rectangle();
-  Rectangle(const point_t& pos, const float height, const float width);
+  Rectangle(const point_t& pos, const double height, const double width);
 
-  float       getArea() const noexcept override;
+  double       getArea() const noexcept override;
   rectangle_t getFrameRect() const noexcept override;
   void        printData(std::ostream& stream) const override;
 
   void move(const point_t& point) noexcept override;
-  void move(const float   dx, const float dy) noexcept override;
+  void move(const double   dx, const double dy) noexcept override;
 
 private:
-  float       height_;
-  float       width_;
+  double       height_;
+  double       width_;
   rectangle_t frame_rect_;
 };
