+++ rodchenkov.simon/A1/shape.hpp
@@ -7,12 +7,12 @@
 class Shape
 {
 public:
-  virtual float       getArea() const noexcept = 0;
+  virtual double       getArea() const noexcept = 0;
   virtual rectangle_t getFrameRect() const noexcept = 0;
   virtual void        printData(std::ostream& stream) const = 0;
 
   virtual void move(const point_t& point) noexcept = 0;
-  virtual void move(const float   dx, const float dy) noexcept = 0;
+  virtual void move(const double   dx, const double dy) noexcept = 0;
 
   virtual ~Shape() = default;
 };
